+++ frontend/src/components/Users/DeleteUserModal.jsx
@@ -8,10 +8,7 @@ import DialogContentText from '@mui/material/DialogContentText';
 import DialogTitle from '@mui/material/DialogTitle';
 
 import {useEffect, useRef, useState} from "react";
-import {Paper, Stack, TextField} from "@mui/material";
 import {makeStyles} from "@mui/styles";
-import dispatcherService from "../../services/dispatcher.service";
-import Dispatcher from "../../screens/Dispatcher";
 import DispatcherService from "../../services/dispatcher.service";
 
 const style = {
@@ -76,34 +73,50 @@ function DeleteUserModal(props) {
     const open = props.open
     const handleOpen = props.handleOpen;
     const handleClose = props.handleClose;
+
     const [loading, setLoading] = useState(false)
+    const [isError, setIsError] = useState(false);
+    const [message, setMessage] = useState('');
 
     let rowsSelected = props.selectedRows;
-    let res = "";
 
     //console.log(props.selectedRows.size);
 
-    /*if(props.selectedRows.size != 0) {
-        rowsSelected = props.selectedRows;
-        res = Array.from(rowsSelected).join(' ');
-        console.log(res)
-    }*/
     function DeleteHandler() {
         console.log("entered the handler")
-        for (const element of rowsSelected) {
-            console.log(element);
-            DispatcherService.deleteUsers(element)
-                .then(function (response) {
-                    reload()
-                    console.log(response);
-                })
-                .catch((error) => {
-                    console.log(error)
-                })
+        setLoading(true);
 
+        if (rowsSelected.length === 0) {
+            setIsError(true)
+            setMessage('No rows selected')
+            setLoading(false)
         }
+        else {
+            try {
+                for (const element of rowsSelected) {
+                    //console.log(element);
+                    DispatcherService.deleteUsers(element)
+                        .then(function (response) {
+                            setIsError(false)
+                            setLoading(false)
+                            //console.log(response);
+                        })
+                        .catch((error) => {
+                            setIsError(true)
+                            //setMessage(error.message)
+                            setLoading(false)
+                        })
+                }
+            }
+            catch (e) {
+                console.error(e);
+                setIsError(true)
+                setMessage("Delete not possible")
+                setLoading(false)
+            }
+        }
+        reload();
         handleClose();
-        //update table function
     }
 
     return (
@@ -120,6 +133,13 @@ function DeleteUserModal(props) {
                 <DialogContentText id="alert-dialog-description">
                     You are going to delete {rowsSelected.length} users
                 </DialogContentText>
+                {isError && (
+                    <div className='form-group'>
+                        <div className='alert alert-danger' role='alert'>
+                            {message}
+                        </div>
+                    </div>
+                )}
             </DialogContent>
             <DialogActions>
                 <Button onClick={handleClose} autoFocus variant='contained' color='primary'>No</Button>
